proteus: Add convenient Space-m hotkey to look up man pages
It matches inside word, then runs :sh to open man with that selection as
the first argument. It works decently well, it pops up a helix native
scrollable popup thing that shows the manpage contents. The scrolling in
that type of popup is mildly awkward (C-[u,d]), but it's much more
convenient than manually opening the manpage in a separate shell window.