proteus: .bin/pshot: Use better heuristics for determining session type
pshot is a script in my $PATH that quickly lets me select a region to
capture a screenshot, and publish it to my webserver.

Some arch update at some point broke something, which means the
XDG_SESSION_TYPE env var can no longer be used to determine what kind of
session is running (I use this both on wayland and Xorg).