proteus: Add xterm config
I like how fast xterm is and I use a bitmap font now anyway, so I wanted
to see how close I could get it to xfce4-terminal, and the answer is,
pretty close! The only remaining minor annoyance is the lack of special
symbols for a few things, but I'm using it for now.